requirements.txt
Django>=5.0
djangorestframework>=3.14
psycopg2-binary>=2.9.10
python-dotenv>=1.0
